Move utils file out of integration directory
